requirements.txt
requests==2.31.0
betterproto==2.0.0b6
websocket-client==1.7.0
PyExecJS==1.5.1
mini_racer==0.12.4
python-dotenv==1.0.0